ARTEMIS-986 fix int overflow
(cherry picked from commit 4e92b63de735a74e9156c53ef50ac72b205d59ad)

+++ artemis-jdbc-store/src/main/java/org/apache/activemq/artemis/jdbc/store/drivers/derby/DerbySQLProvider.java
@@ -35,7 +35,7 @@ public class DerbySQLProvider extends GenericSQLProvider {
    }
 
    @Override
-   public int getMaxBlobSize() {
+   public long getMaxBlobSize() {
       return MAX_BLOB_SIZE;
    }
